requirements.txt
Django==4.2.2
python-dotenv==1.0.1
psycopg2==2.9.10
djangorestframework==3.15.2
django-filter==24.3
djangorestframework-simplejwt==5.3.1
coverage==7.6.4
drf-yasg==1.21.8
django-cors-headers==4.6.0
